Audit fixes: performance, security, and UX improvements
Performance: pre-allocate audio buffers in ChunkAssembler and dictation
path to avoid repeated reallocations; reuse AVAudioPCMBuffer across tap
callbacks; batch WAV sample writes into single fwrite.

Security: restrict temp file permissions to 0600 for transcription
privacy; validate output directory writability before accepting; fix
narrowing size_t→int cast in queue bounds check.

UX: remove stale README references (Translate to English, French/German);
add macOS notification on recording complete; add About dialog with
version and model info; show recording sub-states (Listening/Capturing);
prompt restart when selecting a different model.

+++ src/whisper_worker.cpp
@@ -46,7 +46,7 @@ bool WhisperWorker::start() {
 void WhisperWorker::enqueue(std::vector<float> chunk, int64_t chunk_start_ms,
                             bool is_dictation) {
     std::lock_guard<std::mutex> lock(mutex_);
-    if ((int)queue_.size() >= PROCESSING_QUEUE_MAX) {
+    if (queue_.size() >= static_cast<size_t>(PROCESSING_QUEUE_MAX)) {
         fprintf(stderr, "warn: transcription queue full, dropping oldest chunk\n");
         queue_.pop();
         if (on_error_) on_error_("Transcription queue full — audio chunk dropped");
